backend/requirements.txt
Flask
Flask-Cors
google-auth
google-auth-oauthlib
google-api-python-client
requests
python-dotenv